src/extraction/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.4
murmurhash==1.0.10
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
openai==1.12.0
packaging==23.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.2.0
preshed==3.0.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.1.12
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
